+++ ydb/library/yql/udfs/common/knn/knn.cpp
@@ -1,3 +1,6 @@
+#include "knn-enumerator.h"
+#include "knn-serializer.h"
+
 #include <ydb/library/yql/public/udf/udf_helpers.h>
 
 #include <util/generic/buffer.h>
@@ -6,136 +9,24 @@
 using namespace NYql;
 using namespace NYql::NUdf;
 
-template <typename TCallback>
-void EnumerateVector(const TUnboxedValuePod vector, TCallback&& callback) {
-    const auto elements = vector.GetElements();
-    if (elements) {
-        const auto size = vector.GetListLength();
-        
-        for (ui32 i = 0; i < size; ++i) {
-            callback(elements[i].Get<float>());
-        }
-    } else {
-        TUnboxedValue value;
-        const auto it = vector.GetListIterator();
-        while (it.Next(value)) {
-            callback(value.Get<float>());
-        }
-    }
-}
 
-template <typename TCallback>
-bool EnumerateVectors(const TUnboxedValuePod vector1, const TUnboxedValuePod vector2, TCallback&& callback) {
+SIMPLE_STRICT_UDF(TToBinaryString, char*(TAutoMap<TListType<float>>)) {
+    const TUnboxedValuePod x = args[0];
+
+    const EFormat format = EFormat::FloatVector; // will be taken from args in future
     
-    auto enumerateBothSized = [&callback] (const TUnboxedValuePod vector1, const TUnboxedValue* elements1, const TUnboxedValuePod vector2, const TUnboxedValue* elements2) {
-        const auto size1 = vector1.GetListLength();
-        const auto size2 = vector2.GetListLength();
-        
-        // Length mismatch
-        if (size1 != size2)
-            return false;
+    const ISerializer& serializer = TSerializerFacade::Get(format);
 
-        for (ui32 i = 0; i < size1; ++i) {
-            callback(elements1[i].Get<float>(), elements2[i].Get<float>());
-        }
-        
-        return true;
-    };
-    
-    auto enumerateOneSized = [&callback] (const TUnboxedValuePod vector1, const TUnboxedValue* elements1, const TUnboxedValuePod vector2) {
-        const auto size = vector1.GetListLength();
-        ui32 idx = 0;
-        TUnboxedValue value;
-        const auto it = vector2.GetListIterator();          
-
-        while (it.Next(value)) {
-            callback(elements1[idx++].Get<float>(), value.Get<float>());
-        }
-
-        // Length mismatch
-        if (it.Next(value) || idx != size)
-            return false;
-
-        return true;
-    };
-
-    auto enumerateNoSized = [&callback] (const TUnboxedValuePod vector1, const TUnboxedValuePod vector2) {
-        TUnboxedValue value1, value2;
-        const auto it1 = vector1.GetListIterator();
-        const auto it2 = vector2.GetListIterator();    
-        for (; it1.Next(value1) && it2.Next(value2);) {
-            callback(value1.Get<float>(), value2.Get<float>());
-        }
-
-        // Length mismatch
-        if (it1.Next(value1) || it2.Next(value2))
-            return false;
-
-        return true;
-    };
-
-    const auto elements1 = vector1.GetElements();
-    const auto elements2 = vector2.GetElements();
-    if (elements1 && elements2) {
-        if (!enumerateBothSized(vector1, elements1, vector2, elements2))
-            return false;
-    } else if (elements1) {
-        if (!enumerateOneSized(vector1, elements1, vector2))
-            return false;
-    } else if (elements2) {
-        if (!enumerateOneSized(vector2, elements2, vector1))
-            return false;
-    } else {
-        if (!enumerateNoSized(vector1, vector2))
-            return false;
-    }
-
-    return true;
-}
-
-enum EFormat : unsigned char {
-    FloatVector = 1
-};
-
-TUnboxedValue SerializeFloatVector(const IValueBuilder* valueBuilder, const TUnboxedValuePod x) {
-    const EFormat format = EFormat::FloatVector;
-    
     TString str;
     TStringOutput outStr(str);
 
-    if (x.HasFastListLength()) {
-        const auto size = x.GetListLength();
-        outStr.Reserve(1 + size * sizeof(float));
-    }
+    ui32 size = serializer.CalcSerializedSize(x);
+    if (size)
+        outStr.Reserve(1 + size);
+    
     outStr.Write(&format, sizeof(unsigned char));
 
-    EnumerateVector(x,  [&outStr] (float element) { outStr.Write(&element, sizeof(float)); });
-
-    return valueBuilder->NewString(str);
-}
-
-TUnboxedValue DeserializeFloatVector(const IValueBuilder *valueBuilder, TStringRef str) {
-    if (str.Size() % sizeof(float) != 0)    
-        return {};
-    
-    const ui32 count = str.Size() / sizeof(float);
-
-    TUnboxedValue* items = nullptr;
-    auto res = valueBuilder->NewArray(count, items);
-    
-    TMemoryInput inStr(str);
-    for (ui32 i = 0; i < count; ++i) {
-        float element;
-        if (inStr.Read(&element, sizeof(float)) != sizeof(float))
-            return {};
-        *items++ = TUnboxedValuePod{element};
-    }
-
-    return res.Release();
-}
-
-SIMPLE_STRICT_UDF(TToBinaryString, char*(TAutoMap<TListType<float>>)) {
-    return SerializeFloatVector(valueBuilder, args[0]);
+    return serializer.Serialize(valueBuilder, x, str);
 }
 
 SIMPLE_STRICT_UDF(TFromBinaryString, TOptional<TListType<float>>(const char*)) {
@@ -143,14 +34,12 @@ SIMPLE_STRICT_UDF(TFromBinaryString, TOptional<TListType<float>>(const char*)) {
     if (str.Size() == 0)
         return {};
 
-    char formatByte = str.Data()[0];
+    ui8 formatByte = str.Data()[0];
     str = TStringRef{str.Data() + 1, str.Size() -1};
-    switch (formatByte) {
-        case EFormat::FloatVector: 
-            return DeserializeFloatVector(valueBuilder, str);
-        default:
-            return {};
-    }
+
+    const ISerializer& serializer = TSerializerFacade::Get(formatByte);
+
+    return serializer.Deserialize(valueBuilder, str);
 }
